refactor: add header nav event emits

+++ src/pages/recipe/recipe.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit, OnDestroy, ChangeDetectorRef } from '@angular/core';
-import { NavController, NavParams, Events } from 'ionic-angular';
+import { Component, OnInit, OnDestroy, ViewChildren, QueryList } from '@angular/core';
+import { NavController, NavParams, Events, ItemSliding } from 'ionic-angular';
 
 import { RecipeMaster } from '../../shared/interfaces/recipe-master';
 import { Recipe } from '../../shared/interfaces/recipe';
@@ -19,45 +19,51 @@ import { ToastProvider } from '../../providers/toast/toast';
   templateUrl: 'recipe.html'
 })
 export class RecipePage implements OnInit, OnDestroy {
-  private recipeTabState = 'outOfView';
-  private masterList: Array<RecipeMaster> = null;
-  private masterRecipeList: Array<Recipe> = null;
-  private isLoggedIn: boolean = false;
-  private hasActiveBatch: boolean = false;
-  private masterIndex: number = -1;
-  private creationMode: boolean = false;
-  private _userUpdate: any;
-  private _tabChange: any;
-  private _newMaster: any;
-  private _updateRecipe: any;
+  @ViewChildren('slidingItems') slidingItems: QueryList<ItemSliding>;
+  masterList: Array<RecipeMaster> = null;
+  masterRecipeList: Array<Recipe> = null;
+  isLoggedIn: boolean = false;
+  hasActiveBatch: boolean = false;
+  masterIndex: number = -1;
+  creationMode: boolean = false;
+  _userUpdate: any;
+  _newMaster: any;
+  _updateRecipe: any;
 
   constructor(public navCtrl: NavController,
     public navParams: NavParams,
-    private events: Events,
-    private cdRef: ChangeDetectorRef,
-    private userService: UserProvider,
-    private recipeService: RecipeProvider,
-    private toastService: ToastProvider) {
+    public events: Events,
+    public userService: UserProvider,
+    public recipeService: RecipeProvider,
+    public toastService: ToastProvider) {
       this._userUpdate = this.userUpdateEventHandler.bind(this);
-      this._tabChange = this.tabChangeEventHandler.bind(this);
       this._newMaster = this.newMasterEventHandler.bind(this);
       this._updateRecipe = this.updateRecipeEventHandler.bind(this);
   }
 
   deleteMaster(master: RecipeMaster) {
-    this.recipeService.deleteRecipeMasterById(master._id)
-      .subscribe(response => {
-        console.log('deleted master', response);
-        const index = getIndexById(master._id, this.masterList);
-        if (index != -1) {
-          this.masterList.splice(index, 1);
-          this.toastService.presentToast('Recipe master deleted!', 1500);
-        }
-      });
+    console.log('delete master', master);
+    if (master.hasActiveBatch) {
+      console.log('cannot delete');
+      // TODO toast cannot delete a master with active batch
+    } else {
+      this.recipeService.deleteRecipeMasterById(master._id)
+        .subscribe(response => {
+          const index = getIndexById(master._id, this.masterList);
+          if (index !== -1) {
+            this.masterList.splice(index, 1);
+            this.toastService.presentToast('Recipe master deleted!', 1500);
+          }
+        });
+    }
   }
 
   expandMaster(index: number) {
-    this.masterIndex = this.masterIndex == index ? -1: index;
+    this.masterIndex = this.masterIndex === index ? -1: index;
+  }
+
+  ionViewDidLeave() {
+    this.slidingItems.forEach(slidingItem => slidingItem.close());
   }
 
   getMasterList() {
@@ -71,14 +77,22 @@ export class RecipePage implements OnInit, OnDestroy {
   mapMasterRecipes() {
     this.masterRecipeList = this.masterList.map(master => {
       const selected =  master.recipes.find(recipe => {
-        return recipe._id == master.master;
+        return recipe._id === master.master;
       });
-      return selected == undefined ? master.recipes[0]: selected;
+      return selected === undefined ? master.recipes[0]: selected;
     });
   }
 
   navToBrewProcess(master: RecipeMaster) {
-    if (this.recipeService.isRecipeProcessPresent(master.recipes.find(recipe => recipe._id == master.master))) {
+    if (this.recipeService.isRecipeProcessPresent(
+      master.recipes.find(recipe => recipe._id === master.master)
+    )) {
+      this.events.publish('header-nav-update', {
+        dest: 'process',
+        destType: 'page',
+        destTitle: master.recipes.find(recipe => recipe._id === master.master).variantName,
+        origin: this.navCtrl.getActive().name
+      });
       this.navCtrl.push(ProcessPage, {
         master: master,
         requestedUserId: master.owner,
@@ -90,10 +104,22 @@ export class RecipePage implements OnInit, OnDestroy {
   }
 
   navToDetails(index: number) {
+    this.events.publish('header-nav-update', {
+      dest: 'process',
+      destType: 'page',
+      destTitle: this.masterList[index].name,
+      origin: this.navCtrl.getActive().name
+    });
     this.navCtrl.push(RecipeMasterDetailPage, {master: this.masterList[index]});
   }
 
   navToRecipeForm() {
+    this.events.publish('header-nav-update', {
+      dest: 'recipe-form',
+      destType: 'page',
+      destTitle: 'Create Recipe',
+      origin: this.navCtrl.getActive().name
+    });
     this.navCtrl.push(RecipeFormPage, {formType: 'master', mode: 'create'});
   }
 
@@ -106,20 +132,20 @@ export class RecipePage implements OnInit, OnDestroy {
       // TODO check and pull from storage
     }
     this.events.subscribe('user-update', this._userUpdate);
-    this.events.subscribe('tab-change', this._tabChange);
     this.events.subscribe('new-master', this._newMaster);
     this.events.subscribe('update-recipe', this._updateRecipe);
   }
 
   ngOnDestroy() {
     this.events.unsubscribe('user-update', this._userUpdate);
-    this.events.unsubscribe('tab-change', this._tabChange);
     this.events.unsubscribe('new-master', this._newMaster);
     this.events.unsubscribe('update-recipe', this._updateRecipe);
   }
 
   showExpandedMaster(index: number): boolean {
-    return index == this.masterIndex;
+    return index === this.masterIndex && this.masterList[index].recipes.some(
+      recipe => recipe.processSchedule.length > 0
+    );
   }
 
   toggleCreationMode() {
@@ -136,17 +162,13 @@ export class RecipePage implements OnInit, OnDestroy {
     this.getMasterList();
   }
 
-  tabChangeEventHandler(tab: any): void {
-    this.recipeTabState = tab.dest == 'recipe' ? 'inView': 'outOfView';
-  }
-
   updateRecipeEventHandler(recipe: Recipe) {
     const recipeMaster = this.masterList.find(master => {
       return master.recipes.some(_recipe => {
-        return _recipe._id == recipe._id;
+        return _recipe._id === recipe._id;
       });
     });
-    const indexToUpdate = recipeMaster.recipes.findIndex(item => item._id == recipe._id);
+    const indexToUpdate = recipeMaster.recipes.findIndex(item => item._id === recipe._id);
     recipeMaster.recipes[indexToUpdate] = recipe;
     recipeMaster.master = recipe._id;
     this.mapMasterRecipes();
